--- src/modules/index.ts
@@ -1,3 +1,0 @@
-export * from './prisma/prisma.module';
-export * from './prisma/prisma.service';
-export * from './points/points.module';
